+++ Classes/Command/PageLanguageOverlayCommandController.php
@@ -4,8 +4,12 @@ namespace T3v\T3vDataMapper\Command;
 use TYPO3\CMS\Core\Database\QueryGenerator;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
-use T3v\T3vCore\Command\AbstractCommandController;
+use LucidFrame\Console\ConsoleTable;
 
+use T3v\T3vCore\Command\AbstractCommandController;
+use T3v\T3vCore\Utility\StringUtility;
+
+use T3v\T3vDataMapper\Domain\Model\Page;
 use T3v\T3vDataMapper\Domain\Model\Page\LanguageOverlay;
 use T3v\T3vDataMapper\Service\DatabaseService;
 
@@ -32,9 +36,9 @@ class PageLanguageOverlayCommandController extends AbstractCommandController {
   protected $databaseService;
 
   /**
-   * The list command.
+   * Lists page language overlays.
    *
-   * @param int $sysLanguageUid The system language UID to list
+   * @param int $sysLanguageUid The system language UID of the page language overlays to list
    * @param int $pid The optional PID of the page to search from, defaults to `1`
    * @param int $recursion The optional recursion, defaults to `99`
    * @param string $exclude The optional UIDs of pages to exclude from as string, seperated by `,`, empty by default
@@ -42,29 +46,73 @@ class PageLanguageOverlayCommandController extends AbstractCommandController {
   public function listCommand(int $sysLanguageUid, int $pid = 1, int $recursion = 99, string $exclude = '') {
     $this->beforeCommand();
 
-    $exclude       = $exclude === '-' ? '' : $exclude;
-    $exclude       = GeneralUtility::intExplode(',', $exclude, true);
-    $pagesTreeList = $this->queryGenerator->getTreeList($pid, $recursion, 0, 1);
-    $pageUids      = GeneralUtility::intExplode(',', $pagesTreeList, true);
-    $pageUids      = array_diff($pageUids, $exclude);
+    $table = new ConsoleTable();
+    $table->setHeaders(['UID', 'Title', 'Page ID', 'Status']);
 
-    foreach ($pageUids as $pageUid) {
+    foreach ($this->getPageUids($pid, $recursion, $exclude) as $pageUid) {
       $languageOverlay = LanguageOverlay::where([['pid', '=', $pageUid], ['sys_language_uid', '=', $sysLanguageUid]])->first();
 
       if ($languageOverlay) {
         $uid    = $languageOverlay->uid;
-        $title  = $languageOverlay->title;
+        $title  = StringUtility::asciify($languageOverlay->title);
         $status = $languageOverlay->hidden ? 'hidden' : 'visible';
 
-        $this->log("{$title} ({$uid}) is {$status}", 'white', true);
+        $table->addRow([$uid, $title, $pageUid, $status]);
+      }
+    }
+
+    $table->display();
+  }
+
+  /**
+   * Initializes page language overlays.
+   *
+   * @param int $sysLanguageUid The system language UID of the page language overlays to initialize
+   * @param int $pid The optional PID of the page to search from, defaults to `1`
+   * @param int $recursion The optional recursion, defaults to `99`
+   * @param string $exclude The optional UIDs of pages to exclude from as string, seperated by `,`, empty by default
+   * @param bool $verbose The optional verbosity, defaults to `false`
+   */
+  public function initializeCommand(int $sysLanguageUid, int $pid = 1, int $recursion = 99, string $exclude = '', bool $verbose = false) {
+    $this->beforeCommand();
+
+    foreach ($this->getPageUids($pid, $recursion, $exclude) as $pageUid) {
+      $languageOverlay = LanguageOverlay::where([['pid', '=', $pageUid], ['sys_language_uid', '=', $sysLanguageUid]])->first();
+
+      if ($languageOverlay) {
+        $uid   = $languageOverlay->uid;
+        $title = $languageOverlay->title;
+
+        $this->log("Language overlay `{$title}` ({$uid}) for the page with UID {$pageUid} already exists, skipping...", 'warning', $verbose);
+      } else {
+        $page = Page::where([['uid', '=', $pageUid]])->first();
+
+        $languageOverlay                     = new LanguageOverlay;
+        $languageOverlay->pid                = $pageUid;
+        $languageOverlay->sys_language_uid   = $sysLanguageUid;
+        $languageOverlay->doktype            = $page->doktype;
+        $languageOverlay->title              = $page->title;
+        $languageOverlay->nav_title          = $page->nav_title;
+        $languageOverlay->title              = $page->title;
+        $languageOverlay->keywords           = $page->keywords;
+        $languageOverlay->description        = $page->description;
+        $languageOverlay->abstract           = $page->abstract;
+        $languageOverlay->author             = $page->author;
+        $languageOverlay->author_email       = $page->author_email;
+        $languageOverlay->tx_t3vpage_summary = $page->tx_t3vpage_summary;
+        $languageOverlay->tx_t3vpage_claim   = $page->tx_t3vpage_claim;
+        $languageOverlay->tx_t3vpage_outline = $page->tx_t3vpage_outline;
+        $languageOverlay->save();
+
+        $this->log("New page language overlay for the page with UID {$pageUid} initialized.", 'ok', $verbose);
       }
     }
   }
 
   /**
-   * The hide command.
+   * Hides page language overlays.
    *
-   * @param int $sysLanguageUid The system language UID to hide
+   * @param int $sysLanguageUid The system language UID of the page language overlays to hide
    * @param int $pid The optional PID of the page to search from, defaults to `1`
    * @param int $recursion The optional recursion, defaults to `99`
    * @param string $exclude The optional UIDs of pages to exclude from as string, seperated by `,`, empty by default
@@ -73,13 +121,7 @@ class PageLanguageOverlayCommandController extends AbstractCommandController {
   public function hideCommand(int $sysLanguageUid, int $pid = 1, int $recursion = 99, string $exclude = '', bool $verbose = false) {
     $this->beforeCommand();
 
-    $exclude       = $exclude === '-' ? '' : $exclude;
-    $exclude       = GeneralUtility::intExplode(',', $exclude, true);
-    $pagesTreeList = $this->queryGenerator->getTreeList($pid, $recursion, 0, 1);
-    $pageUids      = GeneralUtility::intExplode(',', $pagesTreeList, true);
-    $pageUids      = array_diff($pageUids, $exclude);
-
-    foreach ($pageUids as $pageUid) {
+    foreach ($this->getPageUids($pid, $recursion, $exclude) as $pageUid) {
       $languageOverlay = LanguageOverlay::where([['pid', '=', $pageUid], ['sys_language_uid', '=', $sysLanguageUid]])->first();
 
       if ($languageOverlay) {
@@ -88,23 +130,21 @@ class PageLanguageOverlayCommandController extends AbstractCommandController {
         $hidden = $languageOverlay->hidden;
 
         if (!$hidden) {
-          $this->log("Hiding {$title} ({$uid})...", 'info', $verbose);
-
           $languageOverlay->hidden = true;
           $languageOverlay->save();
 
-          $this->log("{$title} ({$uid}) is now hidden.", 'ok', $verbose);
+          $this->log("Language overlay `{$title}` ({$uid}) of the page with UID {$pageUid} is now hidden.", 'ok', $verbose);
         } else {
-          $this->log("{$title} ({$uid}) is already hidden, skipping...", 'warning', $verbose);
+          $this->log("Language overlay `{$title}` ({$uid}) of the page with UID {$pageUid} is already hidden, skipping...", 'warning', $verbose);
         }
       }
     }
   }
 
   /**
-   * The unhide command.
+   * Unhides page language overlays.
    *
-   * @param int $sysLanguageUid The system language UID to unhide
+   * @param int $sysLanguageUid The system language UID of the page language overlays to unhide
    * @param int $pid The optional PID of the page to search from, defaults to `1`
    * @param int $recursion The optional recursion, defaults to `99`
    * @param string $exclude The optional UIDs of pages to exclude from as string, seperated by `,`, empty by default
@@ -113,13 +153,7 @@ class PageLanguageOverlayCommandController extends AbstractCommandController {
   public function unhideCommand(int $sysLanguageUid, int $pid = 1, int $recursion = 99, string $exclude = '', bool $verbose = false) {
     $this->beforeCommand();
 
-    $exclude       = $exclude === '-' ? '' : $exclude;
-    $exclude       = GeneralUtility::intExplode(',', $exclude, true);
-    $pagesTreeList = $this->queryGenerator->getTreeList($pid, $recursion, 0, 1);
-    $pageUids      = GeneralUtility::intExplode(',', $pagesTreeList, true);
-    $pageUids      = array_diff($pageUids, $exclude);
-
-    foreach ($pageUids as $pageUid) {
+    foreach ($this->getPageUids($pid, $recursion, $exclude) as $pageUid) {
       $languageOverlay = LanguageOverlay::where([['pid', '=', $pageUid], ['sys_language_uid', '=', $sysLanguageUid]])->first();
 
       if ($languageOverlay) {
@@ -128,14 +162,12 @@ class PageLanguageOverlayCommandController extends AbstractCommandController {
         $hidden = $languageOverlay->hidden;
 
         if ($hidden) {
-          $this->log("Unhiding {$title} ({$uid})...", 'info', $verbose);
-
           $languageOverlay->hidden = false;
           $languageOverlay->save();
 
-          $this->log("{$title} ({$uid}) is now visible.", 'ok', $verbose);
+          $this->log("Language overlay `{$title}` ({$uid}) of the page with UID {$pageUid} is now visible.", 'ok', $verbose);
         } else {
-          $this->log("{$title} ({$uid}) is already visible, skipping...", 'warning', $verbose);
+          $this->log("Language overlay `{$title}` ({$uid}) of the page with UID {$pageUid} is already visible, skipping...", 'warning', $verbose);
         }
       }
     }
@@ -146,5 +178,22 @@ class PageLanguageOverlayCommandController extends AbstractCommandController {
    */
   protected function beforeCommand() {
     $this->databaseService->setup();
+  }
+
+  /**
+   * Gets page UIDs.
+   *
+   * @param int $pid The optional PID of the page to search from, defaults to `1`
+   * @param int $recursion The optional recursion, defaults to `99`
+   * @param string $exclude The optional UIDs of pages to exclude from as string, seperated by `,`, empty by default
+   */
+  protected function getPageUids(int $pid = 1, int $recursion = 99, string $exclude = '') {
+    $exclude       = $exclude === '-' ? '' : $exclude;
+    $exclude       = GeneralUtility::intExplode(',', $exclude, true);
+    $pagesTreeList = $this->queryGenerator->getTreeList($pid, $recursion, 0, 1);
+    $pageUids      = GeneralUtility::intExplode(',', $pagesTreeList, true);
+    $pageUids      = array_diff($pageUids, $exclude);
+
+    return $pageUids;
   }
 }