+++ ext_localconf.php
@@ -10,5 +10,6 @@ $extensionKey = $_EXTKEY;
 if (TYPO3_MODE === 'BE') {
   // --- Extbase Command Controllers ---
 
-  $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['extbase']['commandControllers'][$extensionKey] = \T3v\T3vDataMapper\Command\PageLanguageOverlayCommandController::class;
+  $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['extbase']['commandControllers'][$extensionKey] =
+    \T3v\T3vDataMapper\Command\PageLanguageOverlayCommandController::class;
 }