+++ CHANGELOG.md
@@ -3,6 +3,16 @@ CHANGELOG
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Improved `AbstractModel` and `PageLanguageOverlayCommandController`
+* Introduced `strict` and `fallback` mode (breaking change)
+* Moved functionality from `PageService` to `ExtensionService`
+* Updated dependencies
+* Updated Travis CI and AppVeyor configuration
+* Cleaned up
+
 4.8.0
 -----